+++ .angular/cache/17.3.4/vite/deps/@angular_core.js
@@ -451,7 +451,7 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-ORK62JUO.js";
+} from "./chunk-O2WIL5MP.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
