+++ .angular/cache/17.3.4/vite/deps/@angular_router.js
@@ -1,15 +1,13 @@
 import {
-  Title
-} from "./chunk-RWB2BZ5T.js";
-import {
   DOCUMENT,
   HashLocationStrategy,
   LOCATION_INITIALIZED,
   Location,
   LocationStrategy,
   PathLocationStrategy,
+  Title,
   ViewportScroller
-} from "./chunk-YRGNH7G3.js";
+} from "./chunk-HGFPX6GE.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   APP_INITIALIZER,
@@ -112,7 +110,7 @@ import {
   ɵɵloadQuery,
   ɵɵqueryRefresh,
   ɵɵsanitizeUrlOrResourceUrl
-} from "./chunk-ORK62JUO.js";
+} from "./chunk-O2WIL5MP.js";
 
 // node_modules/@angular/router/fesm2022/router.mjs
 var PRIMARY_OUTLET = "primary";
