+++ src/app/app.component.ts
@@ -4,7 +4,7 @@ import { Component, OnInit } from '@angular/core';
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
-  standalone: true // Agregar esta línea
+  standalone: true
 })
 export class AppComponent implements OnInit {
   title = 'CaronteCHK';
